+++ lib/src/domain/entity/list_wind.dart
@@ -1,6 +1,6 @@
 class ListWind {
   final double speed;
-  final int degrees;
+  final double degrees;
   final double gust;
   ListWind({
     required this.speed,
